Move protobuf conversion method from CelValue to CelProtoWrapper in order to continue de-coupling protobuf-specific code from base CelValue.
PiperOrigin-RevId: 331633792
