+++ eval/public/BUILD
@@ -24,13 +24,10 @@ cc_library(
     ],
     deps = [
         ":cel_value_internal",
-        "//internal:proto_util",
-        "@com_google_absl//absl/container:node_hash_map",
         "@com_google_absl//absl/status",
         "@com_google_absl//absl/status:statusor",
         "@com_google_absl//absl/strings",
         "@com_google_absl//absl/strings:str_format",
-        "@com_google_absl//absl/synchronization",
         "@com_google_absl//absl/time",
         "@com_google_absl//absl/types:optional",
         "@com_google_protobuf//:protobuf",
@@ -141,6 +138,7 @@ cc_library(
     deps = [
         ":cel_function",
         ":cel_function_registry",
+        "//eval/public/structs:cel_proto_wrapper",
         "@com_google_absl//absl/status",
         "@com_google_absl//absl/status:statusor",
         "@com_google_absl//absl/strings",
@@ -308,6 +306,7 @@ cc_library(
     ],
     deps = [
         ":cel_value",
+        "//internal:proto_util",
         "@com_google_absl//absl/status",
         "@com_google_absl//absl/strings",
         "@com_google_protobuf//:protobuf",
@@ -339,14 +338,10 @@ cc_test(
         ":unknown_attribute_set",
         ":unknown_set",
         "//base:status_macros",
-        "//eval/public/structs:cel_proto_wrapper",
-        "//eval/testutil:test_message_cc_proto",
-        "//testutil:util",
         "@com_google_absl//absl/status",
         "@com_google_absl//absl/strings",
         "@com_google_absl//absl/time",
         "@com_google_googletest//:gtest_main",
-        "@com_google_protobuf//:protobuf",
     ],
 )
 
@@ -359,6 +354,7 @@ cc_test(
     deps = [
         ":cel_attribute",
         ":cel_value",
+        "//eval/public/structs:cel_proto_wrapper",
         "@com_google_absl//absl/strings",
         "@com_google_googletest//:gtest_main",
         "@com_google_protobuf//:protobuf",
