+++ eval/public/value_export_util.cc
@@ -4,6 +4,7 @@
 #include "google/protobuf/util/time_util.h"
 #include "absl/strings/escaping.h"
 #include "absl/strings/str_cat.h"
+#include "internal/proto_util.h"
 
 namespace google {
 namespace api {
